tools/metrics_collector/pyproject.toml
[project]
name = "geovito-metrics-collector"
version = "0.1.0"
description = "Local-first aggregated metrics collector for GeoVito"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.8.2",
  "requests>=2.32.3",
  "google-auth>=2.35.0",
  "google-auth-oauthlib>=1.2.1",
  "google-api-python-client>=2.151.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[project.scripts]
geovito-metrics-collector = "geovito_metrics_collector.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
